requirements.txt
numpy
scipy
matplotlib
seaborn
pyyaml
astropy
requests
h5py
progressbar
termcolor
fitsio
cython
pymc


